requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
tensorflow==2.17.0
numpy==1.26.4
pillow==10.4.0
httpx==0.25.0
python-multipart==0.0.19